shell.nix: squash missing cert error
CI jobs were failing with "SSL certificate problem: unable to get local
issuer certificate" when performing git operations in the `yarn install`
step for uniswap-v2-core.

This is because CI jobs are running in a pure shell, and the missing
cacert package meant that the root ca certs were not available.

Assuming that this was introduced by the upgrade to nixos 20.03 in klab.